Move DepositorStats into the Statistics module.
